passport-photo-ai/backend/tests/test_heic_conversion_real.py
"""
Real-world HEIC conversion tests.
Tests Requirements 1.2 from the specification.

Note: These tests require actual HEIC files for full validation.
HEIC files cannot be easily generated programmatically, so some tests
use mocking while others provide instructions for manual testing.
"""

import pytest
import os
from unittest.mock import Mock, patch, MagicMock
from PIL import Image
import tempfile


class TestHEICConversionRealWorld:
    """Test HEIC conversion with real-world scenarios."""
    
    def test_convert_heic_function_signature(self):
        """Verify the convert_heic_to_jpeg function has correct signature."""
        from application import convert_heic_to_jpeg
        import inspect
        
        sig = inspect.signature(convert_heic_to_jpeg)
        params = list(sig.parameters.keys())
        
        # Should accept one parameter: heic_file_path
        assert len(params) == 1
        assert 'heic_file_path' in params or 'path' in str(params).lower()
    
    def test_heic_conversion_with_invalid_file(self):
        """Test HEIC conversion with a file that doesn't exist."""
        from application import convert_heic_to_jpeg
        
        # Try to convert a non-existent file
        result = convert_heic_to_jpeg('/nonexistent/path/to/file.heic')
        
        # Should return None on error
        assert result is None
    
    def test_heic_conversion_with_non_heic_file(self):
        """Test HEIC conversion with a JPEG file (should fail gracefully)."""
        from application import convert_heic_to_jpeg
        
        # Create a temporary JPEG file
        with tempfile.NamedTemporaryFile(suffix='.jpg', delete=False) as tmp:
            img = Image.new('RGB', (100, 100), color='red')
            img.save(tmp.name, 'JPEG')
            tmp_path = tmp.name
        
        try:
            # Try to convert a JPEG as if it were HEIC (should fail)
            result = convert_heic_to_jpeg(tmp_path)
            
            # Should return None on error
            assert result is None
        finally:
            # Cleanup
            if os.path.exists(tmp_path):
                os.remove(tmp_path)
    
    def test_heic_output_path_format(self):
        """Test that HEIC conversion generates correct output path."""
        # Test the path transformation logic
        test_cases = [
            ('/path/to/photo.heic', '/path/to/photo.jpg'),
            ('/path/to/photo.HEIC', '/path/to/photo.jpg'),
            ('photo.heic', 'photo.jpg'),
            ('/path/with.dots/photo.heic', '/path/with.dots/photo.jpg'),
        ]
        
        for input_path, expected_output in test_cases:
            # Simulate the path transformation
            output_path = input_path.rsplit('.', 1)[0] + '.jpg'
            assert output_path == expected_output, f"Failed for {input_path}"
    
    def test_heic_conversion_cleanup_on_success(self):
        """Test that original HEIC file handling is correct."""
        from application import convert_heic_to_jpeg
        
        with tempfile.TemporaryDirectory() as tmpdir:
            heic_path = os.path.join(tmpdir, 'test.heic')
            
            # Mock successful conversion
            mock_heif_file = MagicMock()
            mock_heif_file.mode = 'RGB'
            mock_heif_file.size = (800, 600)
            mock_heif_file.data = b'\x00' * (800 * 600 * 3)
            mock_heif_file.stride = 800 * 3
            
            with patch('application.pyheif.read', return_value=mock_heif_file):
                with patch('application.Image.frombytes') as mock_frombytes:
                    mock_image = MagicMock()
                    mock_frombytes.return_value = mock_image
                    
                    result = convert_heic_to_jpeg(heic_path)
                    
                    # Verify conversion was attempted
                    assert result is not None
                    assert result.endswith('.jpg')


class TestHEICConversionDocumentation:
    """Document HEIC conversion testing requirements."""
    
    def test_manual_testing_instructions(self):
        """
        Document manual testing procedure for HEIC conversion.
        
        MANUAL TESTING INSTRUCTIONS:
        ============================
        
        To fully test HEIC conversion with a real HEIC file:
        
        1. Obtain a HEIC file:
           - Take a photo with an iPhone (iOS 11+) in HEIC format
           - Or download a sample HEIC file from the internet
           - Save it as: tests/fixtures/test_real.heic
        
        2. Run the manual test:
           ```python
           from application import convert_heic_to_jpeg
           import os
           
           heic_path = 'tests/fixtures/test_real.heic'
           if os.path.exists(heic_path):
               result = convert_heic_to_jpeg(heic_path)
               print(f"Conversion result: {result}")
               
               if result and os.path.exists(result):
                   print("✓ HEIC conversion successful!")
                   print(f"✓ Output file: {result}")
                   
                   # Verify it's a valid JPEG
                   from PIL import Image
                   img = Image.open(result)
                   print(f"✓ Image size: {img.size}")
                   print(f"✓ Image mode: {img.mode}")
               else:
                   print("✗ HEIC conversion failed")
           else:
               print(f"HEIC test file not found: {heic_path}")
           ```
        
        3. Expected results:
           - A .jpg file should be created in the same directory
           - The JPEG should have the same dimensions as the original
           - The JPEG should be viewable in any image viewer
        
        4. Cleanup:
           - Remove the generated .jpg file after testing
        """
        # This test always passes - it's documentation
        assert True
    
    def test_heic_conversion_requirements(self):
        """
        Document HEIC conversion requirements.
        
        REQUIREMENTS:
        =============
        
        1. pyheif library must be installed (✓ checked in requirements.txt)
        2. libheif system library must be available
        3. On macOS: Install via `brew install libheif`
        4. On Linux: Install via `apt-get install libheif-dev`
        5. On Windows: May require additional setup
        
        SUPPORTED FORMATS:
        ==================
        - Input: .heic, .HEIC (case-insensitive)
        - Output: .jpg (JPEG format)
        
        ERROR HANDLING:
        ===============
        - Returns None if conversion fails
        - Prints error message to console
        - Does not raise exceptions (graceful degradation)
        """
        # Verify pyheif is importable
        try:
            import pyheif
            assert True, "pyheif is available"
        except ImportError:
            pytest.fail("pyheif is not installed")


class TestHEICConversionIntegration:
    """Test HEIC conversion integration with the workflow."""
    
    def test_workflow_handles_heic_extension(self):
        """Test that the workflow correctly identifies HEIC files."""
        heic_filenames = [
            'photo.heic',
            'image.HEIC',
            'Picture.Heic',
            'test.HeiC'
        ]
        
        for filename in heic_filenames:
            is_heic = filename.lower().endswith('.heic')
            assert is_heic is True, f"{filename} should be identified as HEIC"
    
    def test_workflow_skips_non_heic_files(self):
        """Test that non-HEIC files are not processed as HEIC."""
        non_heic_filenames = [
            'photo.jpg',
            'image.jpeg',
            'picture.png',
            'file.gif',
            'document.pdf',
            'heic.txt'  # Contains 'heic' but wrong extension
        ]
        
        for filename in non_heic_filenames:
            is_heic = filename.lower().endswith('.heic')
            assert is_heic is False, f"{filename} should not be identified as HEIC"
    
    def test_converted_jpeg_path_used_in_workflow(self):
        """Test that converted JPEG path replaces original HEIC path."""
        from application import convert_heic_to_jpeg
        
        with tempfile.TemporaryDirectory() as tmpdir:
            original_heic = os.path.join(tmpdir, 'original.heic')
            
            # Mock conversion
            mock_heif_file = MagicMock()
            mock_heif_file.mode = 'RGB'
            mock_heif_file.size = (800, 600)
            mock_heif_file.data = b'\x00' * (800 * 600 * 3)
            mock_heif_file.stride = 800 * 3
            
            with patch('application.pyheif.read', return_value=mock_heif_file):
                with patch('application.Image.frombytes') as mock_frombytes:
                    mock_image = MagicMock()
                    mock_frombytes.return_value = mock_image
                    
                    converted_path = convert_heic_to_jpeg(original_heic)
                    
                    # Verify the workflow would use the converted path
                    assert converted_path != original_heic
                    assert converted_path.endswith('.jpg')
                    
                    # In the workflow, this path would replace the original
                    temp_path = converted_path  # This is what the workflow would use
                    assert temp_path.endswith('.jpg')


def test_heic_library_availability():
    """Test that required HEIC libraries are available."""
    try:
        import pyheif
        print(f"\n✓ pyheif version: {pyheif.libheif_version()}")
        assert True
    except ImportError:
        pytest.fail("pyheif library is not installed")
    except Exception as e:
        print(f"\n⚠ pyheif is installed but may have issues: {e}")
        # Don't fail - library is installed even if there are runtime issues
        assert True
