passport-photo-ai/backend/requirements.txt
flask
flask_cors
Pillow
opencv-python
numpy
python-dotenv
pillow-heif
rembg
gunicorn
anthropic
pytest
hypothesis
pytest-mock
pytest-cov